+++ src/workspace/defaultToolbar.tsx
@@ -1,25 +1,33 @@
 import * as React from 'react';
 import classnames from 'classnames';
+import { saveAs } from 'file-saver';
 
 import { useObservedProperty } from '../coreUtils/hooks';
 
-import { CanvasContext } from '../diagram/canvasApi';
+import { CanvasContext, ExportRasterOptions } from '../diagram/canvasApi';
 import { defineCanvasWidget } from '../diagram/canvasWidget';
-import { CommandHistory } from '../diagram/history';
 import { layoutForcePadded } from '../diagram/layout';
-import { DiagramModel } from '../diagram/model';
-import { DiagramView } from '../diagram/view';
+import { dataURLToBlob } from '../diagram/toSvg';
 
 import { AuthoringState } from '../editor/authoringState';
-import { EditorController } from '../editor/editorController';
 
-import { HamburgerMenu, HamburgerMenuItem } from '../widgets/hamburgerMenu';
+import { Dropdown, DropdownItem, useInsideDropdown } from '../widgets/dropdown';
 
 import { WorkspaceContext } from './workspaceContext';
 
 export interface DefaultToolbarProps {
-    onSaveDiagram?: () => void;
-    onPersistChanges?: () => void;
+    /**
+     * Main menu content, in a form of `<ToolbarItem>` elements.
+     *
+     * If `null`, the menu toggle button will be hidden.
+     */
+    menu?: React.ReactNode | null;
+    /**
+     * Toolbar panel content, in a form of `<ToolbarItem>` or other elements.
+     *
+     * If `null`, the panel will be hidden.
+     */
+    children?: React.ReactNode | null;
     /**
      * Set of languages to display diagram data.
      */
@@ -34,227 +42,304 @@ export interface WorkspaceLanguage {
 const CLASS_NAME = 'reactodia-toolbar';
 
 export function DefaultToolbar(props: DefaultToolbarProps) {
-    const {
-        onSaveDiagram,
-        onPersistChanges,
-        languages = [],
-    } = props;
-    const {
-        model, view, editor,
-        performLayout, exportSvg, exportPng, print,
-    } = React.useContext(WorkspaceContext)!;
-    const {canvas} = React.useContext(CanvasContext)!;
-
-    const [cancellation] = React.useState(() => new AbortController());
-    React.useEffect(() => {
-        return () => cancellation.abort();
-    }, []);
-
+    const {menu, children, languages = []} = props;
+    const menuContent = menu === null ? null : (
+        menu ?? <>
+            <ToolbarActionClearAll />
+            <ToolbarActionExport kind='exportRaster' />
+            <ToolbarActionExport kind='exportSvg' />
+            <ToolbarActionExport kind='print' />
+        </>
+    );
+    const childrenContent = children === null ? null : (
+        children ?? <>
+            <ToolbarActionUndo />
+            <ToolbarActionRedo />
+            <ToolbarActionLayout />
+            <ToolbarLanguageSelector languages={languages} />
+        </>
+    );
     return (
         <div className={CLASS_NAME}>
-            <HamburgerMenu className={`${CLASS_NAME}__menu`}
-                title='Open main menu'>
-                <SaveButtons editor={editor}
-                    onSaveDiagram={onSaveDiagram}
-                    onPersistChanges={onPersistChanges}
-                />
-                <HamburgerMenuItem
-                    className={classnames(
-                        `${CLASS_NAME}__clear-all-button`,
-                        'reactodia-btn reactodia-btn-default'
-                    )}
-                    title='Remove all elements and links from the diagram'
-                    onClick={() => {
-                        const batch = model.history.startBatch('Clear all');
-                        editor.removeItems([...model.elements]);
-                        batch.store();
-                    }}>
-                    Clear All
-                </HamburgerMenuItem>
-                <HamburgerMenuItem
-                    className={classnames(
-                        `${CLASS_NAME}__export-button`,
-                        'reactodia-btn reactodia-btn-default'
-                    )}
-                    title='Export the diagram as a PNG image'
-                    onClick={() => exportPng(canvas)}>
-                    Export as PNG
-                </HamburgerMenuItem>
-                <HamburgerMenuItem
-                    className={classnames(
-                        `${CLASS_NAME}__export-button`,
-                        'reactodia-btn reactodia-btn-default'
-                    )}
-                    title='Export the diagram as an SVG image'
-                    onClick={() => exportSvg(canvas)}>
-                    Export as SVG
-                </HamburgerMenuItem>
-                <HamburgerMenuItem
-                    className={classnames(
-                        `${CLASS_NAME}__print-button`,
-                        'reactodia-btn reactodia-btn-default'
-                    )}
-                    title='Print the diagram'
-                    onClick={() => print(canvas)}>
-                    Print
-                </HamburgerMenuItem>
-            </HamburgerMenu>
-            <div className={`${CLASS_NAME}__quick-access-group reactodia-btn-group reactodia-btn-group-sm`}>
-                <HistoryButtons history={model.history} />
-                <LayoutButton model={model}
-                    onClick={() => performLayout({
-                        canvas,
-                        layoutFunction: layoutForcePadded,
-                        animate: true,
-                        signal: cancellation.signal,
-                    })}
-                />
-                <LanguageSelector view={view}
-                    languages={languages}
-                />
-            </div>
+            {menuContent ? (
+                <Dropdown className={`${CLASS_NAME}__menu`}
+                    title='Open menu'>
+                    {menuContent}
+                </Dropdown>
+            ) : null}
+            {childrenContent ? (
+                <div className={`${CLASS_NAME}__quick-access-group reactodia-btn-group reactodia-btn-group-sm`}>
+                    {childrenContent}
+                </div>
+            ) : null}
         </div>
     );
 }
 
 defineCanvasWidget(DefaultToolbar, element => ({element, attachment: 'viewport'}));
 
-function SaveButtonsRaw(props: {
-    editor: EditorController;
-    onSaveDiagram?: () => void;
-    onPersistChanges?: () => void;
-}) {
-    const {editor, onSaveDiagram, onPersistChanges} = props;
+export interface ToolbarItemProps {
+    className?: string;
+    title?: string;
+    disabled?: boolean;
+    onSelect?: () => void;
+    children?: React.ReactNode;
+}
+
+export function ToolbarItem(props: ToolbarItemProps) {
+    const {className, title, disabled, onSelect, children} = props;
+    const insideDropdown = useInsideDropdown();
+    return insideDropdown ? (
+        <DropdownItem className={className}
+            title={title}
+            disabled={disabled}
+            onSelect={onSelect}>
+            {children}
+        </DropdownItem>
+    ) : (
+        <button type='button'
+            className={classnames(
+                className,
+                'reactodia-btn reactodia-btn-default'
+            )}
+            title={title}
+            disabled={disabled}
+            onClick={onSelect}>
+            {children}
+        </button>
+    );
+}
+
+export interface ToolbarActionSaveProps {
+    mode: 'layout' | 'authoring';
+    onSelect: () => void;
+    children?: React.ReactNode;
+}
+
+export function ToolbarActionSave(props: ToolbarActionSaveProps) {
+    const {mode, onSelect, children} = props;
+    const {model, editor} = React.useContext(WorkspaceContext)!;
+    const hasLayoutChanges = useObservedProperty(
+        model.history.events,
+        'historyChanged',
+        () => model.history.undoStack.length > 0
+    );
     const canPersistChanges = useObservedProperty(
         editor.events,
         'changeAuthoringState',
         () => !AuthoringState.isEmpty(editor.authoringState)
     );
-    const canSaveDiagram = !canPersistChanges;
+
+    let enabled = true;
+    if (mode === 'layout') {
+        enabled = hasLayoutChanges && !canPersistChanges;
+    } else if (mode === 'authoring') {
+        enabled = canPersistChanges;
+    }
 
     return (
-        <>
-            {onSaveDiagram ? (
-                <HamburgerMenuItem
-                    className={classnames(
-                        `${CLASS_NAME}__save-button`,
-                        'reactodia-btn reactodia-btn-default'
-                    )}
-                    disabled={!canSaveDiagram}
-                    onClick={() => onSaveDiagram()}>
-                    Save diagram
-                </HamburgerMenuItem>
-            ) : null}
-            {onPersistChanges ? (
-                <HamburgerMenuItem
-                    className={classnames(
-                        `${CLASS_NAME}__save-button`,
-                        'reactodia-btn reactodia-btn-default'
-                    )}
-                    disabled={!canPersistChanges}
-                    onClick={() => onPersistChanges()}>
-                    Save data
-                </HamburgerMenuItem>
-            ) : null}
-        </>
+        <DropdownItem
+            className={`${CLASS_NAME}__save`}
+            disabled={!enabled}
+            onSelect={onSelect}>
+            {children}
+        </DropdownItem>
     );
 }
 
-const SaveButtons = React.memo(
-    SaveButtonsRaw,
-    (prevProps, nextProps) => !(
-        nextProps.editor === prevProps.editor &&
-        nextProps.onPersistChanges === prevProps.onPersistChanges &&
-        nextProps.onSaveDiagram === prevProps.onSaveDiagram
-    )
-);
+export interface ToolbarActionClearAllProps {}
 
-function HistoryButtonsRaw(props: {
-    history: CommandHistory;
-}) {
-    const {history} = props;
+export function ToolbarActionClearAll(props: ToolbarActionClearAllProps) {
+    const {model, editor} = React.useContext(WorkspaceContext)!;
+    return (
+        <ToolbarItem
+            className={`${CLASS_NAME}__clear-all`}
+            title='Remove all elements and links from the diagram'
+            onSelect={() => {
+                const batch = model.history.startBatch('Clear all');
+                editor.removeItems([...model.elements]);
+                batch.store();
+            }}>
+            Clear All
+        </ToolbarItem>
+    );
+}
 
-    const {undoCommand, redoCommand} = useObservedProperty(
+export interface ToolbarActionExportProps {
+    kind: 'exportRaster' | 'exportSvg' | 'print';
+    /**
+     * Exported file name without extension.
+     *
+     * @default "diagram"
+     */
+    fileName?: string;
+    /**
+     * Exported image background color for raster images.
+     *
+     * @default {backgroundColor: 'white'}
+     */
+    rasterOptions?: ExportRasterOptions;
+}
+
+export function ToolbarActionExport(props: ToolbarActionExportProps) {
+    const {canvas} = React.useContext(CanvasContext)!;
+    const {kind, fileName = 'diagram', rasterOptions} = props;
+    if (kind === 'exportRaster') {
+        return (
+            <ToolbarItem
+                className={`${CLASS_NAME}__export-image`}
+                title='Export the diagram as a PNG image'
+                onSelect={() => {
+                    const exportOptions: ExportRasterOptions = rasterOptions ?? {
+                        backgroundColor: 'white',
+                    };
+                    canvas.exportRaster(exportOptions).then(dataUri => {
+                        const blob = dataURLToBlob(dataUri);
+                        saveAs(blob, `${fileName}.png`);
+                    });
+                }}>
+                Export as PNG
+            </ToolbarItem>
+        );
+    } else if (kind === 'exportSvg') {
+        return (
+            <ToolbarItem
+                className={`${CLASS_NAME}__export-image`}
+                title='Export the diagram as an SVG image'
+                onSelect={() => {
+                    canvas.exportSvg({addXmlHeader: true}).then(svg => {
+                        const blob = new Blob([svg], {type: 'image/svg+xml'});
+                        saveAs(blob, `${fileName}.svg`);
+                    });
+                }}>
+                Export as SVG
+            </ToolbarItem>
+        );
+    } else if (kind === 'print') {
+        return (
+            <ToolbarItem
+                className={`${CLASS_NAME}__print`}
+                title='Print the diagram'
+                onSelect={() => {
+                    const printWindow = window.open('', undefined, 'width=1280,height=720')!;
+                    canvas.exportSvg().then(svg => {
+                        printWindow.document.write(svg);
+                        printWindow.document.close();
+                        printWindow.print();
+                    });
+                }}>
+                Print
+            </ToolbarItem>
+        );
+    } else {
+        return null;
+    }
+}
+
+export interface ToolbarActionUndoProps {}
+
+function ToolbarActionUndoRaw(props: ToolbarActionUndoProps) {
+    const {model: {history}} = React.useContext(CanvasContext)!;
+    const insideDropdown = useInsideDropdown();
+    const undoCommand = useObservedProperty(
         history.events,
         'historyChanged',
         () => {
-            const {undoStack, redoStack} = history;
-            return {
-                undoCommand: undoStack.length === 0
-                    ? undefined : undoStack[undoStack.length - 1],
-                redoCommand: history.redoStack.length === 0
-                    ? undefined : redoStack[redoStack.length - 1],
-            };
-        },
-        ({undoCommand, redoCommand}) => [undoCommand, redoCommand] as const
+            const {undoStack} = history;
+            return undoStack.length === 0
+                ? undefined : undoStack[undoStack.length - 1];
+        }
     );
-
     return (
-        <>
-            <button type='button'
-                className={classnames(
-                    `${CLASS_NAME}__undo-button`,
-                    'reactodia-btn reactodia-btn-default'
-                )}
-                disabled={!undoCommand}
-                title={
-                    undoCommand && undoCommand.title
-                        ? `Undo: ${undoCommand.title}`
-                        : 'Undo last command'
-                }
-                onClick={() => history.undo()}
-            />
-            <button type='button'
-                className={classnames(
-                    `${CLASS_NAME}__redo-button`,
-                    'reactodia-btn reactodia-btn-default'
-                )}
-                disabled={!redoCommand}
-                title={
-                    redoCommand && redoCommand.title
-                        ? `Redo: ${redoCommand.title}`
-                        : 'Redo last command'
-                }
-                onClick={() => history.redo()}
-            />
-        </>
+        <ToolbarItem className={`${CLASS_NAME}__undo`}
+            disabled={!undoCommand}
+            title={
+                undoCommand && undoCommand.title
+                    ? `Undo: ${undoCommand.title}`
+                    : 'Undo last command'
+            }
+            onSelect={() => history.undo()}>
+            {insideDropdown ? 'Undo' : null}
+        </ToolbarItem>
     );
 }
 
-const HistoryButtons = React.memo(
-    HistoryButtonsRaw,
-    (prevProps, nextProps) => nextProps.history !== prevProps.history
+export const ToolbarActionUndo = React.memo(
+    ToolbarActionUndoRaw,
+    (prevProps, nextProps) => true
 );
 
-function LayoutButton(props: {
-    model: DiagramModel;
-    onClick: () => void;
-}) {
-    const {model, onClick} = props;
+export interface ToolbarActionRedoProps {}
+
+function ToolbarActionRedoRaw(props: ToolbarActionRedoProps) {
+    const {model: {history}} = React.useContext(CanvasContext)!;
+    const insideDropdown = useInsideDropdown();
+    const redoCommand = useObservedProperty(
+        history.events,
+        'historyChanged',
+        () => {
+            const {redoStack} = history;
+            return history.redoStack.length === 0
+                ? undefined : redoStack[redoStack.length - 1];
+        }
+    );
+    return (
+        <ToolbarItem className={`${CLASS_NAME}__redo`}
+            disabled={!redoCommand}
+            title={
+                redoCommand && redoCommand.title
+                    ? `Redo: ${redoCommand.title}`
+                    : 'Redo last command'
+            }
+            onSelect={() => history.redo()}>
+            {insideDropdown ? 'Redo' : null}
+        </ToolbarItem>
+    );
+}
+
+export const ToolbarActionRedo = React.memo(
+    ToolbarActionRedoRaw,
+    (prevProps, nextProps) => true
+);
+
+export interface ToolbarActionLayoutProps {}
+
+function ToolbarActionLayoutRaw(props: ToolbarActionLayoutProps) {
+    const {performLayout} = React.useContext(WorkspaceContext)!;
+    const {model, canvas} = React.useContext(CanvasContext)!;
     const elementCount = useObservedProperty(
         model.events,
         'changeCells',
         () => model.elements.length
     );
     return (
-        <button type='button'
-            className={classnames(
-                `${CLASS_NAME}__layout-button`,
-                'reactodia-btn reactodia-btn-default'
-            )}
+        <ToolbarItem className={`${CLASS_NAME}__layout`}
             title='Layout diagram using force-directed algorithm'
             disabled={elementCount === 0}
-            onClick={onClick}>
+            onSelect={() => {
+                performLayout({
+                    canvas,
+                    layoutFunction: layoutForcePadded,
+                    animate: true,
+                });
+            }}>
             Layout
-        </button>
+        </ToolbarItem>
     );
 }
 
-function LanguageSelector(props: {
-    view: DiagramView;
+export const ToolbarActionLayout = React.memo(
+    ToolbarActionLayoutRaw,
+    (prevProps, nextProps) => true
+);
+
+export interface ToolbarLanguageSelectorProps {
     languages: ReadonlyArray<WorkspaceLanguage>;
-}) {
-    const {view, languages} = props;
+}
+
+export function ToolbarLanguageSelector(props: ToolbarLanguageSelectorProps) {
+    const {view} = React.useContext(WorkspaceContext)!;
+    const {languages} = props;
     const currentLanguage = useObservedProperty(
         view.events,
         'changeLanguage',